Upgrade port selection UI
