+++ twnzui/__init__.py
@@ -1,4 +1,4 @@
 from .port_selector import PortSelectionGUI
 from .login_form import LoginApplication, K_RESULT
-
+from . import frame
 # credit for icon https://icon-icons.com/icon/random-line/72612
