+++ twnzui/port_selector.py
@@ -1,43 +1,24 @@
-from PyQt5.QtWidgets import QApplication, QMainWindow, QWidget, QLabel, QPushButton, QVBoxLayout
+from PyQt5.QtWidgets import QPushButton, QVBoxLayout, QSizePolicy
 
-class PortSelectionGUI(QMainWindow):
+from twnzui.frame import NostyFrame
+
+class PortSelectionGUI(NostyFrame):
     def __init__(self, ports: [tuple], out_port: list):
-        super().__init__()
         self.ports = ports
         self.out_port = out_port
-        self.button_labels = ['|'.join(p) for p in self.ports]
-        self.setWindowTitle("Twnz Nosty Run")
+        NostyFrame.__init__(self, "Nosty Bot - Select Client", 150)
 
-        # Create central widget
-        self.central_widget = QWidget(self)
-        self.setCentralWidget(self.central_widget)
+    def create_elems(self):
+        self.button_labels = [' | '.join(p) for p in self.ports]
 
-        # Create a label to display messages
-        self.label = QLabel("", self.central_widget)
-
-        # Create buttons from the list
+    def setup_middle(self, layout_to_add_widget: QVBoxLayout):
         for button_label in self.button_labels:
             button = QPushButton(button_label, self.central_widget)
+            button.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Fixed)
             button.clicked.connect(lambda checked, label=button_label: self.button_click(label))
-
-        # Set layout and positioning
-        self.layout = QVBoxLayout(self.central_widget)
-        self.layout.addWidget(self.label)
-        for button in self.central_widget.findChildren(QPushButton):
-            self.layout.addWidget(button)
-        self.setCentralWidget(self.central_widget)
-
-        # self.setGeometry(0, 0, 400, 300)
-        self.center()
+            layout_to_add_widget.addWidget(button)
 
     def button_click(self, button_text):
         self.out_port.clear()
         self.out_port.append(int(button_text.split("|")[1]))
-        self.label.setText(f"Button {button_text} clicked")
         self.close()
-
-    def center(self):
-        frame_geometry = self.frameGeometry()
-        screen_center = QApplication.desktop().screenGeometry().center()
-        frame_geometry.moveCenter(screen_center)
-        self.move(frame_geometry.topLeft())
